Login and Home page
The login and home pages have been created but both are not yet styled. Routes have been added to allow the user to go to the different pages.

Login
Supabase linked
Users can login using email
Google auth is still needed to allow users to login using google and their email

Home page
Displays  the title and images of the shows. This was fetched using and API